Add files via upload
Hello! Here is what I have so far on the Potter Andresen variables. I wanted to check in and see if anyone has any feedback for me at this stage. I'm also not sure how to approach coding the differential equation for conduction, which I think is the next step.