Add files via upload
This is a test file for implementing pysolar, the library I found that computes solar radiation (link to their documentation here: https://pysolar.readthedocs.io/en/latest/). Yajun asked if it would be easy to use this now that we are working in 1D.

Theoretically I am sure we can use it and it should be a lot more accurate than our current method for finding solar radiation. Right now I'm stuck on some syntax errors so I'm going to set this aside and work on something else for a while. The error I get happens when I try to get solar radiation for an array of multiple datetime objects using Python's builtin datetime library (this way we can compute solar radiation over the needed time interval). I think there is something wrong with the method I'm using to generate the time interval. It's outputting an array of strings instead of datetime objects. So the problem is actually with my use of the datetime library, not the pysolar library. If anyone figures out what's missing, please update!